test: 스토리북 관심사별로 파일 분리
